requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
djangorestframework==3.13.1
drf-spectacular==0.24.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.16.0
packaging==21.3
phonenumbers==8.12.57
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-django==4.5.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
